Sort the wikidata data by unlocode
The order is inconsistent when downloading, and this causes the list to be reordered every time, making working with git inefficient.

In a few cases, there are multiple items with the same unlocode (like http://www.wikidata.org/entity/Q9287543 vs http://www.wikidata.org/entity/Q9287543), so include that in the ordering as well to have a stable order.